chore + fix: App.js added in root path
It seems there's no easy way to make 'react-native-scripts' or 'expo' to
run an initial 'App.js' file which is not located in the root path unless we
modify the actual module.

In order to be able to run the app, I added a new 'App.js' file at the
root path with a component that does nothing but to render the 'App'
component located in the 'src' folder.

Also, the references to non-existant images in 'app.json' have been
removed to avoid warnings when building the app.