feat: Display list of decks stored in Redux
A new 'DeckList' component has been created to display a list of decks
along with the number of questions available in each of them.

The displayed information is fetched from the Redux store, and said store is now
being initialized from the sample json file included in the 'res/data'
folder. Eventually that must change so data is retrieved from the AsyncStorage.