feat: Fetch deck list from AsyncStorage
The action for fetching the deck list is now complete and in use. The
action creator does access the AsyncStorage to fetch the data.

In the reducer, if the AsyncStorage does return no data, we initialize
the store with the json stored in the 'data' folder as we did before.

The 'DeckList' component is using the newly created action to set the
deck list prop when the app starts.