feat: New deck form validation
When submitting the new deck form, the app does now check if the title
field is not empty and, in case it is not, if the introduced name doesn't
match that of an already existing deck.

In either of these two scenarios, an error message will be displayed.
Otherwise, the action to create the deck will be dispatched.

Also, for making it easier to check whether a deck already exists, I've
removed the uuid package and established the actual title of the deck as a
key instead of using uuids to identify each deck.