feat: New screen for adding a new card
The user can now navigate to a create card view. This will contain a form
with four fields: one for writing the question, and three for writing the
possible answers the final user should chose from when taking a quiz.

Each of these answers have a button that can be used to set which one is
the right answer.

The submit button does noting so far: we're neither validating the form
values nor dispatching an action to add the new card to our deck into the
AsyncStorage.