feat: Answer reveal
When the user presses a question in the quiz, the right answer is
highlighted in a different color.

If this feature is used, the user score will remain the same even after
pressing the right answer during that turn. This would count as cheating
because the user saw the answer before pressing on the button, which is
the reason I've decided to not increment the score in this case.