feat: Schedule local notification
A new file has been included in "utils" for handling local
notifications. The file is the same that was created during the course,
for the follow-up project. This is likely to change in upcoming commits.

I've tested that the notification does trigger successfully by trying
to schedule it for 5 minutes from now. The behaviour shall be improved
though: anytime the user completes a quiz, the notification is
re-scheduled. This may mean that the variable in AsyncStorage is not
being set properly and is being overriden all the time.