fix: Redirect to deck details after deck creation
Instead of redirecting to the deck list page, the app was supposed to
redirect the user to the deck details screen for the newly created deck.

Doing this lead for the back button to present an undesired behaviour,
leading back to the create deck form whenever the user pushed the back
button in the details page. Instead, the desired output was to get the
user back into the deck list.

In order to make this happen, the navigation queue is now being reset
before redirecting to the details page, so the create form is no longer
part of the queue and we can go back straight to the list.