refactor: Quiz results added to main navigator
The component to display the quiz results is now part of the application's
main navigator, being considered as a new navigation item rather than an
element rendered in DeckQuiz when a certain condition is met.

This change is not without any drawbacks, though, as the exit button
action needs to be redone. The restart button action has already been
modified to work as expected after the latest changes.