feat: Random order in quiz questions
Since small decks questions are repeated multiple times, the answers are
no longer being displayed in the same fixed order in which they were first
introducing when creating the card.

The fact that the answers are shuffled whenever a new card is displayed
will make the experience a bit more challenging to the final user.

This closes #1.