fix: Don't lose deck title when adding a card
When a new card was added to a deck, the whole deck data was being
removed. The only thing that was preserved was the questions array, but
the title was lost. That caused the deck name not to show up anymore
after the card had been added, leaving it blank when the user got back
to the details page.