fix: Quiz results including right score
The quiz results page wasn't showing the right score. It seems as if the
user was navigating to the results page before the score was updated in
the state, so the max score displayed would always be 90% even if the
user was supposed to get a perfect score.

This commit fixes this issue and closes #4.