refactor: New component for custom text input
Styles and other props common to every input in the app are now included
in a new component.

Also, string refs to focus on a certain input have been replaced by
callbacks as specified in the docs, since those string references are
deprecated and didn't work when we wanted to focus an input inside a
children component.