design: Disable reveal button on quiz
Once the user presses on a card to reveal the answers for the first time,
the flip card functionality won't be triggered again for the same card. Also, a new message will show up on screen to let the user know he already
revealed the truth.

The button's functionality will be re-enabled once the user moves on to
the next question.