chore: Explicit version for react navigation
When trying to clone the project and make a clean install, it seems as if
the latest version of 'react-navigation' caused an issue with the header
height in the StackNavigator.

To prevent any issues like this, the version number is now fixed to the
one I used during development process.