fix: Import react in HomeTabs file
The reviewer came across a reference error when trying to run the application in
iOS, due to 'React' variable being missing somewhere.

I checked all of the component files, and 'HomeTabs' was likely the one to
be causing trouble. The 'tabBarIcon' property uses JSX, which requires
React to be imported into the file. The addition of this import line
should fix the problem described in issue #9.

As far as I remember, the 'tabBarIcon' value was only used in iOS, which
may explain why I never got this error while debugging the app in any of
the devices I used to test (all of them were Android).